backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.2
aiosmtplib==3.0.1
google-generativeai==0.5.4
groq==0.9.0
python-dotenv==1.0.1
pydantic[email]==2.7.1
httpx==0.27.0
